regression: default sorting of dates
The improved comparator function missed a special but
not so uncommon case where the column data consists of dates.
In that case the new comparator also interpreted the data
as strings which leads to a wrong sort order.
Therefore, the comparator now explicitly checks if the type
of the data is string and only then uses the string comparator.
In all other cases, the default object comparator is used which
uses the default natural order of the objects (which is defined
for date objects and most date-like types from various libraries).

Signed-off-by: Mike Becker <Mike.Becker@yasc.de>