fix missing margin between create and cancel
This commit introduces a margin similar to the margin
between the save and cancel button in the tbody.

However, this way of adding margin sucks, because it requires
that at least two children are present. The original developers
did not know that. But for consistency we stick to this for now.

Signed-off-by: Mike Becker <Mike.Becker@yasc.de>